Homework2/requirements.txt
aiogram
openai
python-dotenv
aiofiles
python-docx
PyPDF2
requests
docx2txt
httpx==0.27.2